+++ HISTORY.rst
@@ -10,6 +10,8 @@ Release History
   `Item.download()`, and `File.download()`.
 - `ia download` is now verbose by default. Output can be suppressed with the `--quiet`
   flag.
+- Added an option to not download into item directories, but rather the current working
+  directory (i.e. `ia download --no-directories <id>`).
 
 0.7.1 2014-08-25
 ++++++++++++++++